requirements.txt
streamlit
google-generativeai
extra-streamlit-components
gTTS
dotenv